chore: rename test runner to: custard
