--- ariwete/go.mod
@@ -1,3 +1,0 @@
-module ariwete
-
-go 1.22.0
