+++ custard/go.mod
@@ -1,0 +1,3 @@
+module custard
+
+go 1.22.0
