+++ custard/pkg/config/testdata/config/simple.json
